fix logout in uboat app
