+++ EnigmaMachine/UBoatApplication/src/MainScene/MainUBoatScenePane.fxml
@@ -1,10 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ScrollPane?>
 <?import javafx.scene.control.Tab?>
 <?import javafx.scene.control.TabPane?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.text.Font?>
 
 <ScrollPane fitToHeight="true" fitToWidth="true" minHeight="100.0" minWidth="100.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="MainScene.MainUBoatScenePaneController">
    <content>
@@ -32,6 +35,16 @@
               </tabs>
             </TabPane>
          </center>
+         <bottom>
+            <Button fx:id="logoutButton" mnemonicParsing="false" onAction="#onLogoutButtonClicked" prefHeight="34.0" prefWidth="177.0" text="Logout" textFill="#e80b0b" BorderPane.alignment="CENTER_RIGHT">
+               <BorderPane.margin>
+                  <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+               </BorderPane.margin>
+               <font>
+                  <Font size="18.0" />
+               </font>
+            </Button>
+         </bottom>
       </BorderPane>
    </content>
 </ScrollPane>
